remove redux tools from presentation_utils plugin (#197891)
(cherry picked from commit fb885eaeffcdbbcab399f11ab1e8e524b174bd69)
