+++ src/plugins/dashboard/public/dashboard_api/unified_search_manager.ts
@@ -39,8 +39,8 @@ import {
   connectToQueryState,
   syncGlobalQueryStateWithUrl,
 } from '@kbn/data-plugin/public';
-import { cleanFiltersForSerialize } from '@kbn/presentation-util-plugin/public';
 import moment, { Moment } from 'moment';
+import { cleanFiltersForSerialize } from '../utils/clean_filters_for_serialize';
 import { dataService } from '../services/kibana_services';
 import { DashboardCreationOptions, DashboardState } from './types';
 import { DEFAULT_DASHBOARD_INPUT, GLOBAL_STATE_STORAGE_KEY } from '../dashboard_constants';
