+++ src/plugins/presentation_util/README.mdx
@@ -8,51 +8,4 @@ tags: ['kibana', 'presentation', 'services']
 related: []
 ---
 
-## Introduction
-
 The Presentation Utility Plugin is a set of common, shared components and toolkits for solutions within the Presentation space, (e.g. Dashboards, Canvas).
-
-## Redux Embeddables
-
-The Redux Embeddables system allows embeddable authors to interact with their embeddables in a standardized way using Redux toolkit. This wrapper abstracts away store and slice creation, and embeddable input sync. To use this system, a developer can use CreateReduxEmbeddableTools in the constructor of their embeddable, supplying a collection of reducers.
-
-### Reducers
-
-The reducer object expected by the ReduxEmbeddableWrapper is the same type as the reducers expected by [Redux Toolkit's CreateSlice](https://redux-toolkit.js.org/api/createslice).
-
-<DocAccordion buttonContent="Reducers Example" initialIsOpen>
-```ts
-// my_embeddable_reducers.ts
-import { MyEmbeddableInput } from './my_embeddable';
-
-export const myEmbeddableReducers = {
-  setSpecialBoolean: (
-    state: WritableDraft<MyEmbeddableInput>,
-    action: PayloadAction<MyEmbeddableInput['specialBoolean']>
-  ) => {
-    state.specialBoolean = action.payload;
-  }
-}
-    
-```
-</DocAccordion>
-
-### Accessing Actions and State
-
-From components under the embeddable, actions, containerActions, and the current state of the redux store are accessed via the embeddable instance. You can pass the embeddable instance down as a prop, or use a context.
-
-<DocAccordion buttonContent="Accessing Redux Embeddable Context" initialIsOpen>
-  ```ts
-  // my_embeddable_component.tsx
-  const MyEmbeddableComponent = ({ embeddableInstance }: { embeddableInstance: IEmbeddable }) => {
-    // current state
-    const specialBoolean = embeddableInstance.select((state) => state.specialBoolean);
-
-    // change specialBoolean after 5 seconds
-    setTimeout(() => embeddableInstance.dispatch.setSpecialBoolean(false), 5000);
-
-}
-
-```
-</DocAccordion>
-```
