+++ src/plugins/presentation_util/public/index.ts
@@ -30,15 +30,6 @@ export {
   DEFAULT_DASHBOARD_DRILLDOWN_OPTIONS,
 } from './components';
 
-export {
-  lazyLoadReduxToolsPackage,
-  cleanFiltersForSerialize,
-  type ReduxEmbeddableState,
-  type ReduxEmbeddableTools,
-  type ReduxTools,
-  type ReduxToolsPackage,
-} from './redux_tools';
-
 export type {
   ExpressionInputEditorRef,
   ExpressionInputProps,
